+++ decomp/ST.exe/functions/004A4240/decomp.c
@@ -91,27 +91,27 @@ void __thiscall STGroupBoatC::ChangeMDNotify(STGroupBoatC *this,int param_1,uint
         return;
       }
       STFishC::sub_004162B0((STFishC *)this_01,&local_a,&local_8,&local_6);
-      if (local_a < (short)this_00->field_026A) {
+      if (local_a < this_00->field_026A) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_0270 + -1 + (int)(short)this_00->field_026A < (int)local_a) {
+      if (this_00->field_0270 + -1 + (int)this_00->field_026A < (int)local_a) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (local_8 < (short)this_00->field_026C) {
+      if (local_8 < this_00->field_026C) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_0272 + -1 + (int)(short)this_00->field_026C < (int)local_8) {
+      if (this_00->field_0272 + -1 + (int)this_00->field_026C < (int)local_8) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (local_6 < (short)this_00->field_026E) {
+      if (local_6 < this_00->field_026E) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_0274 + -1 + (int)(short)this_00->field_026E < (int)local_6) {
+      if (this_00->field_0274 + -1 + (int)this_00->field_026E < (int)local_6) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
@@ -169,27 +169,27 @@ void __thiscall STGroupBoatC::ChangeMDNotify(STGroupBoatC *this,int param_1,uint
         return;
       }
       STFishC::sub_004162B0((STFishC *)this_01,&local_a,&local_8,&local_6);
-      if (local_a < (short)this_00->field_0252) {
+      if (local_a < this_00->field_0252) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_0258 + -1 + (int)(short)this_00->field_0252 < (int)local_a) {
+      if (this_00->field_0258 + -1 + (int)this_00->field_0252 < (int)local_a) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (local_8 < (short)this_00->field_0254) {
+      if (local_8 < this_00->field_0254) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_025A + -1 + (int)(short)this_00->field_0254 < (int)local_8) {
+      if (this_00->field_025A + -1 + (int)this_00->field_0254 < (int)local_8) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (local_6 < (short)this_00->field_0256) {
+      if (local_6 < this_00->field_0256) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
-      if (this_00->field_025C + -1 + (int)(short)this_00->field_0256 < (int)local_6) {
+      if (this_00->field_025C + -1 + (int)this_00->field_0256 < (int)local_6) {
         g_currentExceptionFrame = local_78.previous;
         return;
       }
@@ -246,27 +246,27 @@ void __thiscall STGroupBoatC::ChangeMDNotify(STGroupBoatC *this,int param_1,uint
       return;
     }
     STFishC::sub_004162B0((STFishC *)this_01,&local_a,&local_8,&local_6);
-    if (local_a < (short)this_00->field_026A) {
+    if (local_a < this_00->field_026A) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_0270 + -1 + (int)(short)this_00->field_026A < (int)local_a) {
+    if (this_00->field_0270 + -1 + (int)this_00->field_026A < (int)local_a) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (local_8 < (short)this_00->field_026C) {
+    if (local_8 < this_00->field_026C) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_0272 + -1 + (int)(short)this_00->field_026C < (int)local_8) {
+    if (this_00->field_0272 + -1 + (int)this_00->field_026C < (int)local_8) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (local_6 < (short)this_00->field_026E) {
+    if (local_6 < this_00->field_026E) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_0274 + -1 + (int)(short)this_00->field_026E < (int)local_6) {
+    if (this_00->field_0274 + -1 + (int)this_00->field_026E < (int)local_6) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
@@ -347,27 +347,27 @@ void __thiscall STGroupBoatC::ChangeMDNotify(STGroupBoatC *this,int param_1,uint
       return;
     }
     STFishC::sub_004162B0((STFishC *)local_18,&local_a,&local_8,&local_6);
-    if (local_a < (short)this_00->field_0252) {
+    if (local_a < this_00->field_0252) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_0258 + -1 + (int)(short)this_00->field_0252 < (int)local_a) {
+    if (this_00->field_0258 + -1 + (int)this_00->field_0252 < (int)local_a) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (local_8 < (short)this_00->field_0254) {
+    if (local_8 < this_00->field_0254) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_025A + -1 + (int)(short)this_00->field_0254 < (int)local_8) {
+    if (this_00->field_025A + -1 + (int)this_00->field_0254 < (int)local_8) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (local_6 < (short)this_00->field_0256) {
+    if (local_6 < this_00->field_0256) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
-    if (this_00->field_025C + -1 + (int)(short)this_00->field_0256 < (int)local_6) {
+    if (this_00->field_025C + -1 + (int)this_00->field_0256 < (int)local_6) {
       g_currentExceptionFrame = local_78.previous;
       return;
     }
