+++ decomp/ST.exe/functions/00575CB0/decomp.c
@@ -520,7 +520,7 @@ LAB_00576308:
         local_10 = (AnonShape_00575CB0_065D0A66 *)&pAVar5->field_0002;
       } while (local_18 < pSVar18->field_003C);
     }
-    STPlaySystemC::SaveObjData(g_playSystem_00802A38,pSVar18->field_0018,local_1c,local_20);
+    STPlaySystemC::SaveObjData(g_playSystem_00802A38,pSVar18->field_0018,local_1c,(uint)local_20);
     FreeAndNull(&local_1c);
     g_currentExceptionFrame = local_a4.previous;
     return 0;
