+++ src/ST.exe/source/recovered/by_owner/STDcResourcC.cpp
@@ -228,7 +228,7 @@ undefined4 __thiscall st::fn_005805C0(STDcResourcC *this)
       st::fn_00405240((STT3DSprC *)this_00,0xe,g_playSystem_00802A38->field_00E4);
       st::fn_004045D9
                 ((STT3DSprC *)this_00,(float)(int)this->field_0245 * _DAT_007904f8 + _DAT_007904f4,
-                 (float)this->field_0249 * _DAT_007904f8 + _DAT_007904f4,
+                 (float)(int)this->field_0249 * _DAT_007904f8 + _DAT_007904f4,
                  (float)(int)this->field_024D * _DAT_00790504 + _DAT_00790500 + _DAT_007904fc);
       st::fn_006EA3E0((void *)this->field_0211,this->field_01ED,-1);
       st::fn_00403AFD((STT3DSprC *)this_00,'\r');
